Rename 'projectStorage.js' to 'localStorage.js'
To be more in line with the naming scheme of the other clients.
